+++ lib/Patat/Effect.hs
@@ -8,9 +8,14 @@ module Patat.Effect
 
 
 --------------------------------------------------------------------------------
-import           Data.List.NonEmpty (NonEmpty ((:|)))
-import           Data.Unique        (Unique, newUnique)
-import qualified Patat.PrettyPrint  as PP
+import           Data.Foldable              (for_)
+import           Data.List.NonEmpty         (NonEmpty ((:|)))
+import           Data.Unique                (Unique, newUnique)
+import qualified Data.Vector                as V
+import qualified Data.Vector.Mutable        as VM
+import           Patat.Presentation.Display (Size (..))
+import qualified Patat.PrettyPrint          as PP
+import           Patat.PrettyPrint.Matrix
 
 
 --------------------------------------------------------------------------------
@@ -20,21 +25,19 @@ newtype EffectId = EffectId Unique deriving (Eq)
 --------------------------------------------------------------------------------
 data Effect = Effect
     { eId     :: EffectId
-    , eFrames :: NonEmpty PP.Doc
+    , eFrames :: NonEmpty Matrix
     , eDelay  :: Int
     }
 
 
 --------------------------------------------------------------------------------
-newEffect :: IO Effect
-newEffect =
-    Effect <$> (EffectId <$> newUnique) <*> pure frames <*> pure 500000
+newEffect :: Size -> PP.Doc -> PP.Doc -> IO Effect
+newEffect size frame0 frame1 =
+    Effect <$> (EffectId <$> newUnique) <*> pure frames <*> pure 10000
   where
-    frames =
-        PP.string "new slide in 3..." <> PP.hardline :|
-        PP.string "new slide in 2..." <> PP.hardline :
-        PP.string "new slide in 1..." <> PP.hardline :
-        []
+    matrix0 = docToMatrix size frame0
+    matrix1 = docToMatrix size frame1
+    frames  = slide size matrix0 matrix1
 
 
 --------------------------------------------------------------------------------
@@ -43,3 +46,22 @@ stepEffect eid eff | eid /= eId eff = Just eff
 stepEffect _   eff                  = case eFrames eff of
     _ :| []     -> Nothing
     _ :| f : fs -> Just eff {eFrames = f :| fs}
+
+
+--------------------------------------------------------------------------------
+slide :: Size -> Matrix -> Matrix -> NonEmpty Matrix
+slide (Size rows cols) initial final =
+    initial :| map frame [1 .. cols - 1] ++ [final]
+  where
+    frame offset = V.create $ do
+        ini <- V.unsafeThaw initial
+        fin <- V.unsafeThaw final
+        mat <- VM.replicate (rows * cols) $ Elem [] ' '
+        for_ [0 .. rows - 1] $ \y -> do
+            VM.copy
+                (VM.slice (y * cols) (cols - offset) mat)
+                (VM.slice (y * cols + offset) (cols - offset) ini)
+            VM.copy
+                (VM.slice (y * cols + cols - offset) offset mat)
+                (VM.slice (y * cols) offset fin)
+        pure mat
