+++ patat.cabal
@@ -53,6 +53,7 @@ Library
     time                 >= 1.4  && < 1.13,
     unordered-containers >= 0.2  && < 0.3,
     yaml                 >= 0.8  && < 0.12,
+    vector               >= 0.13 && < 0.14,
     wcwidth              >= 0.0  && < 0.1,
     -- We don't even depend on these packages but they can break cabal install
     -- because of the conflicting 'Network.URI' module.
@@ -88,6 +89,8 @@ Library
     Patat.Presentation.Read
     Patat.Presentation.Settings
     Patat.PrettyPrint
+    Patat.PrettyPrint.Internal
+    Patat.PrettyPrint.Matrix
     Patat.Theme
 
   Other-modules:
